style artificial elliptic words in graph
Decrease the `font-size` of `artifical="elliptic"` words to `0.7rem` in
the `<Graph>` component. Due to some issues either with Dagre-d3 or with
SVG rendering in browsers, the solution is slightly more complicated
than setting the `font-size`. See the comment in the `.elliptic` class
in `src/lib/components/Graph/Graph.module.scss`.