move treebank parsing logic to parsing.js
Previously, the logic that converted a the sentence JSON to the format
used by Dagre-d3 was split between Graph.js and parsing.js.

Move all that code from Graph.js to parsing.js so it is in one place.
parsing.js is also a better place for this logic to live because it
won't mix up React component logic with parsing.