fetch config instead of using local config
Instead of including `arethusa-configs` as a Git submodule and
importing the configuration from there, query configs from
`https://arethusa-configs.perseids.org/` (or a user-supplied URL).

The problem with the previous approach is that either the JSON of all the
configurations has to be bundled in the package or dynamic imports have to be
used. Including all of the JSON is doable now, but may become burdensome
in the future if there are hundreds of configurations. Dynamic imports
work locally, but do not work when the code is packaged as a library.

See:

https://stackoverflow.com/questions/62239203/using-dynamic-import-within-a-webpack-exported-library
https://stackoverflow.com/questions/61332255/webpack-async-import-from-one-part-of-a-library-to-another