accept value and setter for active in sentence
In the `<Sentence>` component, accept `active` and `setActive`. `active`
is the `id` of the actively selected word. `setActive` is a function
that accepts a word `id` and makes that word active.

When `active` is present, the `active` word is set to the word with that
`id` even when the sentence `id` or JSON change. If `active` is not
present, the `active` word `id` is set to `null` when the sentence
changes.